pyproject.toml
[tool.poetry]
name = "bot"
version = "1.0.0"
description = "The community bot for the Python Discord community."
authors = ["Python Discord <info@pythondiscord.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "3.9.*"

"discord.py" = {url = "https://github.com/Rapptz/discord.py/archive/beafaa8a8b8006357c2a1ca6802ed80e000f4cda.zip"}
# See https://bot-core.pythondiscord.com/ for docs.
# bot-core = {url = "https://github.com/python-discord/bot-core/archive/refs/tags/v4.0.0.zip"}

aiodns = "3.0.0"
aiohttp = "3.8.1"
aioredis = "1.3.1"
arrow = "1.2.2"
async-rediscache = { version = "0.2.0", extras = ["fakeredis"] }
beautifulsoup4 = "4.10.0"
colorama = { version = "0.4.4", markers = "sys_platform == 'win32'" }
coloredlogs = "15.0.1"
deepdiff = "5.7.0"
emoji = "1.7.0"
feedparser = "6.0.8"
rapidfuzz = "2.0.7"
lxml = "4.8.0"
markdownify = "0.10.3"
more_itertools = "8.12.0"
python-dateutil = "2.8.2"
python-frontmatter = "1.0.0"
pyyaml = "6.0"
regex = "2022.3.15"
sentry-sdk = "1.5.8"
statsd = "3.3.0"
tldextract = "3.2.0"

[tool.poetry.dev-dependencies]
coverage = "6.3.2"
flake8 = "4.0.1"
flake8-annotations = "2.7.0"
flake8-bugbear = "22.1.11"
flake8-docstrings = "1.6.0"
flake8-string-format = "0.3.0"
flake8-tidy-imports = "4.6.0"
flake8-todo = "0.7"
flake8-isort = "4.1.1"
pep8-naming = "0.12.1"
pre-commit = "2.17.0"
taskipy = "1.10.1"
pip-licenses = "3.5.3"
python-dotenv = "0.19.2"
pytest = "7.1.1"
pytest-cov = "3.0.0"
pytest-xdist = "2.5.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.taskipy.tasks]
start = "python -m bot"
lint = "pre-commit run --all-files"
precommit = "pre-commit install"
build = "docker build -t ghcr.io/python-discord/bot:latest -f Dockerfile ."
push = "docker push ghcr.io/python-discord/bot:latest"
test-nocov = "pytest -n 8"
test = "pytest -n 8 --cov-report= --cov --ff"
retest = "pytest -n 8 --cov-report= --cov --lf"
html = "coverage html"
report = "coverage report"
isort = "isort ."

[tool.coverage.run]
branch = true
source_pkgs = ["bot"]
source = ["tests"]

[tool.isort]
multi_line_output = 6
order_by_type = false
case_sensitive = true
combine_as_imports = true
line_length = 120
atomic = true
